Separated redactors from core to accounts. Improved navbar
