+++ accounts/forms.py
@@ -1,0 +1,38 @@
+from django import forms
+from django.contrib.auth import get_user_model
+from django.contrib.auth.forms import UserCreationForm, UserChangeForm
+
+Redactor = get_user_model()
+
+
+class SearchForm(forms.Form):
+    query = forms.CharField(
+        max_length=255,
+        required=False,
+        label="",
+        widget=forms.TextInput(attrs={"placeholder": "Input something..."})
+    )
+
+
+class RedactorCreationForm(UserCreationForm):
+    class Meta(UserCreationForm.Meta):
+        model = Redactor
+        fields = UserCreationForm.Meta.fields + (
+            "first_name",
+            "last_name",
+            "email",
+            "years_of_experience",
+        )
+
+
+class RedactorUpdateForm(UserChangeForm):
+    password = None
+
+    class Meta:
+        model = Redactor
+        fields = (
+            "first_name",
+            "last_name",
+            "email",
+            "years_of_experience",
+        )
