+++ config/urls.py
@@ -21,5 +21,6 @@ from django.urls import path, include
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("accounts/", include("django.contrib.auth.urls")),
+    path("", include("accounts.urls", namespace="accounts")),
     path("", include("core.urls", namespace="core")),
 ]
