+++ core/forms.py
@@ -1,6 +1,5 @@
 from django import forms
 from django.contrib.auth import get_user_model
-from django.contrib.auth.forms import UserCreationForm, UserChangeForm
 
 from core.models import Newspaper, Topic
 
@@ -48,27 +47,3 @@ class SearchForm(forms.Form):
         label="",
         widget=forms.TextInput(attrs={"placeholder": "Input something..."})
     )
-
-
-class RedactorCreationForm(UserCreationForm):
-    class Meta(UserCreationForm.Meta):
-        model = Redactor
-        fields = UserCreationForm.Meta.fields + (
-            "first_name",
-            "last_name",
-            "email",
-            "years_of_experience",
-        )
-
-
-class RedactorUpdateForm(UserChangeForm):
-    password = None
-
-    class Meta:
-        model = Redactor
-        fields = (
-            "first_name",
-            "last_name",
-            "email",
-            "years_of_experience",
-        )
