+++ templates/base.html
@@ -11,8 +11,6 @@
         crossorigin="anonymous">
   <link rel="stylesheet"
         href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.13.1/font/bootstrap-icons.min.css">
-  {#    {% load static %}#}
-  {#    <link rel="stylesheet" href="{% static 'css/styles.css' %}">#}
 </head>
 <body>
 {% block navbar %}
@@ -41,6 +39,10 @@
   </div>
 </div>
 
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.7/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-ndDqU0Gzau9qJ1lfW4pNLlhNTkCfHzAVBReH9diLvGRem5+R9g2FzA8ZGN954O5Q"
+        crossorigin="anonymous">
+</script>
 <script>
 setTimeout(function() {
   document.querySelectorAll('#django-messages .alert').forEach(function(alert) {
