+++ templates/includes/navbar.html
@@ -1,34 +1,52 @@
-<nav class="navbar navbar-expand-lg navbar-light bg-light mb-4">
+<nav class="navbar navbar-expand-lg navbar-light bg-light border-bottom shadow-sm mb-4">
   <div class="container">
-    <a class="navbar-brand" href="{% url 'core:home' %}">Home</a>
-    <div class="collapse navbar-collapse">
-      <ul class="navbar-nav">
 
-        <li class="nav-item"><a class="nav-link" href="{% url 'core:newspaper-list' %}">Newspapers</a></li>
-        <li class="nav-item"><a class="nav-link" href="{% url 'core:topic-list' %}">Topics</a></li>
-        <li class="nav-item"><a class="nav-link" href="{% url 'core:redactor-list' %}">Redactors</a></li>
+    <a class="navbar-brand fw-bold text-primary" href="{% url 'core:home' %}">
+      <i class="bi bi-buildings me-1"></i> InfoAgency
+    </a>
 
-        {% if user.is_authenticated %}
-          <li class="nav-item d-flex align-items-center me-3">
-            <i class="bi bi-person-circle me-1"></i>
-            <span><a href="{% url 'core:redactor-detail' pk=user.pk %}">{{ user.get_full_name }}</a></span>
-          </li>
-          <li class="nav-item">
-            <form action="{% url 'logout' %}" method="post">
-              {% csrf_token %}
-              <button type="submit" class="btn btn-outline-danger btn-sm w-100">
-                <i class="bi bi-box-arrow-right"></i> Logout
-              </button>
-            </form>
-          </li>
-        {% else %}
-          <li class="nav-item">
-            <a href="{% url 'login' %}" class="btn btn-outline-primary btn-sm w-100">
-              <i class="bi bi-box-arrow-in-right"></i> Login
-            </a>
-          </li>
-        {% endif %}
+    <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNavLight"
+            aria-controls="navbarNavLight" aria-expanded="false" aria-label="Toggle navigation">
+      <span class="navbar-toggler-icon"></span>
+    </button>
+
+    <div class="collapse navbar-collapse" id="navbarNavLight">
+      <ul class="navbar-nav me-auto gap-3">
+        <li class="nav-item">
+          <a class="nav-link" href="{% url 'core:newspaper-list' %}">
+            <i class="bi bi-journal-text me-1"></i> Newspapers
+          </a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link" href="{% url 'core:topic-list' %}">
+            <i class="bi bi-bookmarks me-1"></i> Topics
+          </a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link" href="{% url 'accounts:redactor-list' %}">
+            <i class="bi bi-people me-1"></i> Redactors
+          </a>
+        </li>
       </ul>
+
+      <div class="d-flex flex-column flex-lg-row align-items-lg-center gap-2">
+        {% if user.is_authenticated %}
+          <a class="nav-link fw-semibold " href="{% url 'accounts:redactor-detail' pk=user.pk %}">
+            <i class="bi bi-person-circle me-1"></i> {{ user.get_full_name }}
+          </a>
+          <form action="{% url 'logout' %}" method="post" class="m-0">
+            {% csrf_token %}
+            <button type="submit" class="btn btn-outline-danger btn-sm">
+              <i class="bi bi-box-arrow-right"></i> Logout
+            </button>
+          </form>
+        {% else %}
+          <a href="{% url 'login' %}" class="btn btn-outline-primary btn-sm">
+            <i class="bi bi-box-arrow-in-right"></i> Login
+          </a>
+        {% endif %}
+      </div>
+
     </div>
   </div>
 </nav>
