+++ templates/includes/pagination.html
@@ -1,12 +1,11 @@
 {% load query_transform %}
 {% if is_paginated %}
   <ul class="pagination">
-    <!-- First page -->
+
     <li class="page-item {% if page_obj.number == 1 %}disabled{% endif %}">
       <a class="page-link" href="?{% query_transform request page=1 %}">First</a>
     </li>
 
-    <!-- Previous page -->
     {% if page_obj.has_previous %}
       <li class="page-item">
         <a class="page-link" href="?{% query_transform request page=page_obj.previous_page_number %}">Prev</a>
@@ -17,12 +16,10 @@
       </li>
     {% endif %}
 
-    <!-- Current page -->
     <li class="page-item active">
       <span class="page-link">{{ page_obj.number }} of {{ paginator.num_pages }}</span>
     </li>
 
-    <!-- Next page -->
     {% if page_obj.has_next %}
       <li class="page-item">
         <a class="page-link" href="?{% query_transform request page=page_obj.next_page_number %}">Next</a>
@@ -33,7 +30,6 @@
       </li>
     {% endif %}
 
-    <!-- Last page -->
     <li class="page-item {% if page_obj.number == paginator.num_pages %}disabled{% endif %}">
       <a class="page-link" href="?{% query_transform request page=paginator.num_pages %}">Last</a>
     </li>
